Separates into API controller and view controller
