+++ ngrinder-controller/src/main/java/org/ngrinder/operation/cotroller/AnnouncementApiController.java
@@ -2,10 +2,11 @@ package org.ngrinder.operation.cotroller;
 
 import org.ngrinder.operation.service.AnnouncementService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/announcement/api")
+@RequestMapping("/operation/announcement/api")
 public class AnnouncementApiController {
 
 	@Autowired
@@ -14,5 +15,11 @@ public class AnnouncementApiController {
 	@GetMapping("")
 	public String getAnnouncement() {
 		return announcementService.getOne();
+	}
+
+	@PostMapping("")
+	@PreAuthorize("hasAnyRole('A', 'S')")
+	public void save(@RequestParam String content) {
+		announcementService.save(content);
 	}
 }
